requirements.txt
# Core dependencies
crewai>=0.28.0
crewai-tools>=0.1.6
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.0.1

# Vector database
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Session management
redis>=5.0.1

# Environment management
python-dotenv>=1.0.0

# Web framework and socket communication
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0

# Additional utilities
pydantic>=2.0.0


